requirements.txt
python-dotenv
numpy
pandas
matplotlib
statsmodels
scikit-learn
prophet
etna
catboost
optuna
etna[clearml]
etna[statsforecast]
etna[classification]
etna[wandb]
etna[timesfm]
etna[torch]
